+++ src/table/utilities/test-table-properties.js
@@ -77,6 +77,7 @@ export const allSelectedRows = [
 ];
 
 export const tableOptions: ApiQueryOptions = {
+  resultCount: 100,
   offset: 0,
   count: 5,
   searchOptions: [],